requirements.txt
annotated-types==0.6.0
anyio==3.7.1
awscli==1.27.150
bcrypt==4.0.1
botocore==1.29.150
click==8.1.7
colorama==0.4.4
contextvars==2.4
dnspython==2.4.2
docutils==0.16
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.103.2
h11==0.14.0
httptools==0.6.0
idna==3.4
immutables==0.20
jmespath==1.0.1
passlib==1.7.4
pyasn1==0.5.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
# PyYAML==5.4.1
PyYAML==3.10
rsa==4.7.2
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
urllib3==1.26.16
uvicorn==0.23.2
# uvloop==0.18.0
watchfiles==0.21.0
websockets==11.0.3
opencv-python==4.9.0.80
python-multipart==  0.0.9
nibabel
